Bug 443876 - Allow users to turn off strict validation of JavaScript keywords usage when ECMA5 compliant JavaScript are used in their projects
The fix adds the new preference to JavaScript->Validator->Errors/Warnings page: "Strict validation of JavaScript keywords usage" (default value is true -
in order to make it work in the way it did before). ProblemReporter uses this preference value in order to decide if it has to report syntax errors for
keywords or not. The fix includes the editor's reconciler change required to make the preference value change to play immediately after change.

Signed-off-by: vrubezhny <vrubezhny@exadel.com>